utils/make_db/requirements.txt
google-api-python-client==2.33.0
google-auth-oauthlib==0.4.6
numpy==1.21.5
